pyproject.toml
[tool.poetry]
name = "sharknado"
version = "0.1.0"
description = ""
authors = ["K Maximo"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
rich = "^13.5.2"
typer = "^0.9.0"


[tool.poetry.group.dev.dependencies]
blue = "^0.9.1"
isort = "^5.12.0"
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
taskipy = "^1.12.0"
pandas = "^2.0.3"
matplotlib = "^3.7.2"
python-decouple = "^3.8"


[tool.poetry.group.doc.dependencies]
mkdocs-material = "^9.1.21"
mkdocstrings = "^0.22.0"
mkdocstrings-python = "^1.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "."
addopts = "--doctest-modules"

[tool.isort]
profile = "black"
line_length = 79

[tool.taskipy.tasks]
lint = "blue --check --diff . && isort --check --diff ."
docs = "mkdocs serve"
pre_test = "task lint"
test = "pytest -s -x --cov=sharknado -vv"
post_test = "coverage html"
